Renamed to match other components
